+++ src/Symfony/Component/Validator/Tests/Constraints/ValidValidatorTest.php
@@ -39,6 +39,62 @@ class ValidValidatorTest extends TestCase
 
         $this->assertCount(0, $violations);
     }
+
+    public function testCascadedGroupsAreUsedForNestedValidation()
+    {
+        $validatorBuilder = new ValidatorBuilder();
+        $validator = $validatorBuilder->enableAttributeMapping()->getValidator();
+
+        $parent = new ParentWithCascadedGroups();
+        $violations = $validator->validate($parent, null, ['trigger']);
+
+        // Should have 2 violations: one for Default group constraint, one for extra group constraint
+        $this->assertCount(2, $violations);
+    }
+
+    public function testCascadedGroupsWithoutExplicitGroups()
+    {
+        $validatorBuilder = new ValidatorBuilder();
+        $validator = $validatorBuilder->enableAttributeMapping()->getValidator();
+
+        // the natural spelling: no explicit "groups", validated in the Default group
+        $violations = $validator->validate(new ParentWithNaturalCascadedGroups());
+
+        $this->assertCount(2, $violations);
+    }
+
+    public function testEmptyCascadedGroupsSkipTheNestedValidation()
+    {
+        $validatorBuilder = new ValidatorBuilder();
+        $validator = $validatorBuilder->enableAttributeMapping()->getValidator();
+
+        $violations = $validator->validate(new ParentWithEmptyCascadedGroups(), null, ['trigger']);
+
+        $this->assertCount(0, $violations);
+    }
+
+    public function testCascadedGroupsWithAnExplicitConstraint()
+    {
+        $validatorBuilder = new ValidatorBuilder();
+        $validator = $validatorBuilder->enableAttributeMapping()->getValidator();
+
+        $violations = $validator->validate(new ChildWithMultipleGroups(), new Assert\Valid(cascadedGroups: ['Default', 'extra']));
+
+        $this->assertCount(2, $violations);
+    }
+
+    public function testCascadedGroupsDefaultBehaviorPreserved()
+    {
+        $validatorBuilder = new ValidatorBuilder();
+        $validator = $validatorBuilder->enableAttributeMapping()->getValidator();
+
+        // Without cascadedGroups, only the current group should be used
+        $parent = new ParentWithoutCascadedGroups();
+        $violations = $validator->validate($parent, null, ['trigger']);
+
+        // Should have only 1 violation: only the 'trigger' group constraint is validated
+        $this->assertCount(1, $violations);
+    }
 }
 
 class Foo
@@ -67,4 +123,60 @@ class FooBarBaz
 {
     #[Assert\NotBlank(groups: ['nested'])]
     public $foo;
+}
+
+class ParentWithCascadedGroups
+{
+    #[Assert\Valid(groups: ['trigger'], cascadedGroups: ['Default', 'extra'])]
+    public $child;
+
+    public function __construct()
+    {
+        $this->child = new ChildWithMultipleGroups();
+    }
+}
+
+class ParentWithNaturalCascadedGroups
+{
+    #[Assert\Valid(cascadedGroups: ['Default', 'extra'])]
+    public $child;
+
+    public function __construct()
+    {
+        $this->child = new ChildWithMultipleGroups();
+    }
+}
+
+class ParentWithEmptyCascadedGroups
+{
+    #[Assert\Valid(groups: ['trigger'], cascadedGroups: [])]
+    public $child;
+
+    public function __construct()
+    {
+        $this->child = new ChildWithMultipleGroups();
+    }
+}
+
+class ParentWithoutCascadedGroups
+{
+    #[Assert\Valid(groups: ['trigger'])]
+    public $child;
+
+    public function __construct()
+    {
+        $this->child = new ChildWithMultipleGroups();
+    }
+}
+
+class ChildWithMultipleGroups
+{
+    #[Assert\NotBlank]
+    public $defaultField;
+
+    #[Assert\NotBlank(groups: ['extra'])]
+    public $extraField;
+
+    #[Assert\NotBlank(groups: ['trigger'])]
+    public $triggerField;
 }